+++ packages/examples/src/main.ts
@@ -7,12 +7,19 @@ import App from './App.vue'
 import { initGamepad } from 'web-gamepad'
 initGamepad({
   /** 手柄链接时回调 */
-  onConnected: (gamepad: Gamepad) => {},
+  onConnected: (gamepad: Gamepad) => {
+    console.log("🚀 ~ onConnected:", gamepad)
+    
+  },
   /** 手柄断开链接时回调 */
   onDisconnected: (gamepad: Gamepad) => {},
+  onInput:(gamepad: Gamepad) => {
+  console.log("🚀 ~ gamepad:", gamepad)
+
+  },
   axes: {
-    /** 监听手柄摇杆变化精度 */
-    accuracy: 0.001
+    /** 手柄死区半径 */
+    deadZoneRadius: 0.05
   }
 })
 const app = createApp(App).use(ElementPlus)